Refactor separate switch component into nested component
