+++ components/philips_series_2200/switch/power.h
